\name{getCOEFexo}
\alias{getCOEFexo}

\title{All-country LS coefficient estimates.
}

\encoding{latin1}

\description{Extract all-country LS coefficient estimates.
}
\usage{getCOEFexo(out)
}

\arguments{
  \item{out}{A list object of estimation results generated by GVARest().
}

}

\details{Extract all-country LS coefficient estimates.
}


\value{
  \item{coef}{Country-specific coefficient estimates.}

}


\author{Ho Tsung-wu <tsungwu@ntnu.edu.tw>, College of Management, National Taiwan Normal University.
}

\examples{
data("PriceVol")
data("tradeweight1")
data("tradeweightx")
p=2
FLag=2
lag.max=15
type="const"
ic="SC"
weight.matrix=tradeweightx
mainOUTPUT = GVARest(data=PriceVol,p,FLag,lag.max,type,ic,weight.matrix)
#COEF=getCOEFexo(out=mainOUTPUT)
}

